+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## v0.3.0 - 23.08.2020
+
+### Added
+
+* Forget/Reset password functionality
+* Change password functionality
+
+### Changed
+
+* All used `os.environ.get` to `env` from `environs` library
+
 ## v0.2.0 - 22.08.2020
 
 ### Added